+++ Data/Badge.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace PokeCord
+namespace PokeCord.Data
 {
     public class Badge
     {
@@ -12,9 +12,9 @@ namespace PokeCord
         public int Id { get; set; }
         public string? Name { get; set; }
         public string? Description { get; set; }
-        public string? ImageAddress {  get; set; }
+        public string? ImageAddress { get; set; }
         public int BonusPokeballs { get; set; }
-        public List<PokemonData>? GymPokemon {  get; set; }
+        public List<PokemonData>? GymPokemon { get; set; }
     }
 
     public class BadgeManager
@@ -30,15 +30,15 @@ namespace PokeCord
                 if (allBadges.Any(b => b.Id == badge.Id))
                 {
                     // Player has already earned this badge, do not check for it
-                    allBadges.RemoveAll(b => b.Id == badge.Id);                    
+                    allBadges.RemoveAll(b => b.Id == badge.Id);
                 }
             }
             // Check for new acquisitions
-            foreach(Badge badge in allBadges)
+            foreach (Badge badge in allBadges)
             {
                 // Handle non-pokemon specfic badges
                 if (badge.GymPokemon == null)
-                {                    
+                {
                     badge.GymPokemon = new List<PokemonData>();
                 }
                 // Check whether player has caught all pokemon for this badge
