+++ Data/PlayerData.cs
@@ -4,17 +4,17 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace PokeCord
+namespace PokeCord.Data
 {
     public class PlayerData
     {
         public int Version { get; set; } = 3;
-        public ulong UserId {  get; set; }
+        public ulong UserId { get; set; }
         public string UserName { get; set; }
         public int Experience { get; set; }
         public int Pokeballs { get; set; }
         public int PokemonDollars { get; set; } // New!
-        public List<PokemonData> CaughtPokemon {  get; set; }
+        public List<PokemonData> CaughtPokemon { get; set; }
         public List<Badge> EarnedBadges { get; set; } // Fixedish version
 
         public PlayerData()
