+++ Data/Team.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace PokeCord
+namespace PokeCord.Data
 {
     public class Team
     {
